fix(dm-tool): extract toMonsterFileUrl to Electron-free utility module
The test for toMonsterFileUrl failed in CI because the function was
defined in ipc/monsters.ts which imports ipcMain from electron —
Electron is not installed in the headless CI environment.

Move the pure URL-rewrite logic to compendium/image-url.ts (no Electron
imports). monsters.ts now re-imports from there; monsters.test.ts imports
directly from image-url so no Electron runtime is needed.

244/244 tests pass locally; same logic, zero behavior change.
